fix: move main traits to core module and modify gradient and hessian methods to work better at larger values
